+++ styles/globals.css
@@ -27,3 +27,10 @@ html {
 .mapboxgl-popup-content {
     @apply p-2
 }
+.home__orbe {
+	width: 265px;
+	height: 284px;
+	background: linear-gradient(180deg, hsl(0, 0%, 16%) 93%, hsl(0, 0%, 67%) 100%);
+	border-radius: 135px 135px 0 0;
+    align-self: flex-end;
+}
